requirements.txt
jira==3.2.0
PyGithub==1.55
python_dateutil==2.8.2